@@ -32,66 +33,74 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.0.0",
-    "@angular/common": "^7.0.0",
-    "@angular/compiler": "^7.0.0",
-    "@angular/core": "^7.0.0",
-    "@angular/forms": "^7.0.0",
-    "@angular/platform-browser": "^7.0.0",
-    "@angular/platform-browser-dynamic": "^7.0.0",
-    "@angular/router": "^7.0.0",
-    "@nestjs/common": "5.5.0",
-    "@nestjs/core": "5.5.0",
-    "@nestjs/microservices": "^5.7.4",
-    "@nestjs/mongoose": "^5.2.2",
-    "@ngrx/effects": "6.1.2",
-    "@ngrx/router-store": "6.1.2",
-    "@ngrx/store": "6.1.2",
-    "@nrwl/nx": "7.5.1",
-    "angular-oauth2-oidc": "^5.0.2",
-    "bootstrap": "4.2.1",
+    "@angular/animations": "^8.2.14",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
+    "@nestjs/common": "^6.2.4",
+    "@nestjs/core": "^6.2.4",
+    "@nestjs/microservices": "^6.10.7",
+    "@nestjs/mongoose": "^6.1.2",
+    "@nestjs/platform-express": "^6.2.4",
+    "@nestjs/swagger": "^4.0.9",
+    "@ngrx/effects": "8.5.2",
+    "@ngrx/router-store": "8.5.2",
+    "@ngrx/store": "8.5.2",
+    "@nrwl/angular": "8.9.0",
+    "@nrwl/workspace": "8.9.0",
+    "angular-oauth2-oidc": "^8.0.4",
+    "bootstrap": "4.4.1",
     "class-transformer": "^0.2.0",
-    "class-validator": "^0.9.1",
-    "core-js": "^2.5.4",
+    "class-validator": "^0.11.0",
+    "core-js": "^3.4.7",
     "mongoose": "^5.4.18",
-    "rxjs": "~6.4.0",
-    "zone.js": "^0.8.26"
+    "reflect-metadata": "^0.1.12",
+    "rxjs": "~6.5.3",
+    "swagger-ui-express": "^4.1.2",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.11.0",
-    "@angular/cli": "~7.3.0",
-    "@angular/compiler-cli": "^7.0.0",
-    "@angular/language-service": "^7.0.0",
-    "@nestjs/schematics": "5.11.2",
-    "@nestjs/testing": "5.5.0",
-    "@ngrx/store-devtools": "6.1.2",
-    "@nrwl/builders": "7.5.1",
-    "@nrwl/schematics": "7.5.1",
-    "@types/jasmine": "~2.8.6",
+    "@angular-devkit/build-angular": "~0.803.20",
+    "@angular/cli": "~8.3.20",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
+    "@nestjs/schematics": "^6.3.0",
+    "@nestjs/testing": "^6.2.4",
+    "@ngrx/schematics": "8.5.2",
+    "@ngrx/store-devtools": "8.5.2",
+    "@nrwl/jest": "^8.9.0",
+    "@nrwl/nest": "^8.9.0",
+    "@nrwl/node": "^8.9.0",
+    "@nrwl/workspace": "^8.9.0",
+    "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
-    "@types/jest": "^23.0.0",
+    "@types/jest": "24.0.23",
     "@types/mongoose": "^5.3.21",
-    "@types/node": "~8.9.4",
+    "@types/node": "~12.12.14",
     "@types/supertest": "^2.0.7",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-marbles": "0.4.0",
+    "codelyzer": "^5.0.1",
+    "dotenv": "8.2.0",
+    "jasmine-core": "~3.5.0",
+    "jasmine-marbles": "0.6.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "jest": "^23.0.0",
-    "jest-preset-angular": "6.0.1",
-    "karma": "~3.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "ngrx-store-freeze": "0.2.4",
+    "jest": "24.9.0",
+    "jest-preset-angular": "7.1.1",
+    "karma": "~4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
     "npm-run-all": "^4.1.5",
-    "prettier": "1.15.2",
+    "prettier": "1.19.1",
     "protractor": "~5.4.0",
     "rxjs-marbles": "^5.0.0",
-    "supertest": "^3.4.2",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
+    "supertest": "^4.0.2",
+    "ts-node": "~8.5.4",
+    "tslint": "~5.20.1",
+    "typescript": "~3.5.3"
   }
 }
