@@ -1,8 +1,9 @@
-import { Module } from '@nestjs/common';
+import { Logger, Module } from '@nestjs/common';
 
 import { AppController } from './app.controller';
 
 @Module({
-  controllers: [AppController]
+  controllers: [AppController],
+  providers: [Logger]
 })
 export class AppModule {}
