@@ -2,7 +2,7 @@ import { Global, HttpModule, Module } from '@nestjs/common';
 
 @Global()
 @Module({
-  modules: [HttpModule],
+  imports: [HttpModule],
   exports: [HttpModule]
 })
 export class CoreModule {}
