@@ -1,4 +1,4 @@
-import { ExecutionContext, Injectable, Logger, NestInterceptor } from '@nestjs/common';
+import { CallHandler, ExecutionContext, Injectable, Logger, NestInterceptor } from '@nestjs/common';
 import { Observable } from 'rxjs';
 import { tap } from 'rxjs/operators';
 
@@ -7,9 +7,9 @@ export class LoggerInterceptor implements NestInterceptor {
   constructor(private logger: Logger) {
   }
 
-  intercept(context: ExecutionContext, call$: Observable<any>): Observable<any> {
+  intercept(context: ExecutionContext, next: CallHandler): Observable<any> {
     this.logger.log(`Request-URL: ${context.switchToHttp().getRequest<Request>().url}`);
-    return call$.pipe(
+    return next.handle().pipe(
       tap(response => this.logger.log(`Response: ${JSON.stringify(response)}`))
     );
   }
