@@ -4,11 +4,21 @@
  **/
 
 import { NestFactory } from '@nestjs/core';
-
+import { SwaggerModule, DocumentBuilder } from '@nestjs/swagger';
 import { AppModule } from './app/app.module';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
+
+  const options = new DocumentBuilder()
+    .setTitle('Flights example')
+    .setDescription('The flight API description')
+    .setVersion('1.0')
+    .addTag('flights')
+    .build();
+  const document = SwaggerModule.createDocument(app, options);
+  SwaggerModule.setup('api/docs', app, document);
+
   await app.listen(3333, () => {
     console.log('Listening at http://localhost:3333');
   });
